fix: app-state and CLI bug batch
- scroll offsets are usize (u16 wrapped >65535-line docs); stale offsets
  clamped after every rebuild (blank viewport / link-mode slice panic)
- TOC toggle re-lays out for the reduced width instead of truncating
- search matches recompute on tab switch and rebuild; per-span matching so
  count and highlight agree; one lowercasing algorithm everywhere
- slides: frontmatter stripped before splitting; --watch re-splits the deck
- resize events debounced (80ms) before rebuilding
- graphics encode failure shows a notice instead of a blank hole
- --plain: broken pipe exits 0; multiple files all render; directory arg
  refused instead of launching the TUI browser into a pipe
- config keys spacing/toc/frontmatter now apply
- CJK-aware word counts

+++ src/render/mod.rs
@@ -189,12 +189,12 @@ pub fn render_document_with_search(
     frame: &mut Frame,
     area: Rect,
     lines: &[Line<'static>],
-    scroll_offset: u16,
+    scroll_offset: usize,
     _total_lines: usize,
     search: &SearchState,
     t: &theme::Theme,
 ) {
-    let offset = (scroll_offset as usize).min(lines.len());
+    let offset = scroll_offset.min(lines.len());
     let end = (offset + area.height as usize).min(lines.len());
     let visible = &lines[offset..end];
